+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- MSRV is now 1.60.0.
+
+## [0.5.1] - 2023-03-02
+### Fixed
+- Fix a bug on 32-bit platforms that could cause the square root implementation
+  to return an incorrect result.
+- The `sqrt-table` feature now works without `std` and only requires `alloc`.
 
 ## [0.5.0] - 2022-12-06
 ### Added